implmentation done, have to do some testing

+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/page/content/subsection/request/GroupSubSectionRequest.java
@@ -2,7 +2,7 @@ package gov.cdc.nbs.questionbank.page.content.subsection.request;
 
 import java.util.List;
 
-public record GroupSubSectionRequest(long id, String blockName, List<Batch> batches, long repeatingNbr) {
+public record GroupSubSectionRequest(long id, String blockName, List<Batch> batches, int repeatingNbr) {
     public record Batch(long id, char batchTableAppearIndCd, String batchTableHeader, int batchTableColumnWidth) {
     }
